requirements.txt
torch
torchvision
diffusers
transformers
accelerate
pillow